+++ system/opensite.go
@@ -3,16 +3,31 @@ package system
 import (
 	"github.com/ironstar-io/tokaido/services/docker"
 	"github.com/ironstar-io/tokaido/system/goos"
-
-	"log"
 )
 
 // OpenSite - Linux Root executable
-func OpenSite() {
-	httpsPort := docker.LocalPort("haproxy", "8443")
-	if httpsPort == "" {
-		log.Fatal("Unable to obtain the HTTPS port number. The HAProxy container may be broken")
+func OpenSite(args []string) {
+	var p string
+	services := map[string]string{
+		"haproxy": "8443",
+		"adminer": "8080",
+		"mailhog": "8025",
+		"nginx":   "8082",
 	}
 
-	goos.OpenSite("https://localhost:" + httpsPort)
+	if len(args) >= 1 {
+		for _, arg := range args {
+			if len(services[arg]) > 0 {
+				p = docker.LocalPort(arg, services[arg])
+				proto := "http://"
+				if arg == "haproxy" {
+					proto = "https://"
+				}
+				goos.OpenSite(proto + "localhost:" + p)
+			}
+		}
+	} else {
+		p = docker.LocalPort("haproxy", "8443")
+		goos.OpenSite("https://localhost:" + p)
+	}
 }